Move analytics suite init to suite file
